Rename/space searches controller and route
To pluralized to submissions/searches in line
with rails/CRUD conventions
